update phygeonet code
