+++ examples/phygeonet/heat_equation.py
@@ -1,222 +1,17 @@
 from os import path as osp
 
 import hydra
-import matplotlib
 import matplotlib.pyplot as plt
 import numpy as np
 import paddle
-from matplotlib.collections import PatchCollection
-from matplotlib.patches import Polygon
 from omegaconf import DictConfig
+from utils import dfdx
+from utils import dfdy
+from utils import setAxisLabel
+from utils import visualize2D
 
 import ppsci
 from ppsci.utils import logger
-
-
-def setAxisLabel(ax, type):
-    if type == "p":
-        ax.set_xlabel(r"$x$")
-        ax.set_ylabel(r"$y$")
-    elif type == "r":
-        ax.set_xlabel(r"$\xi$")
-        ax.set_ylabel(r"$\eta$")
-    else:
-        raise ValueError("The axis type only can be reference or physical")
-
-
-def gen_e2vcg(x):
-    nelemx = x.shape[1] - 1
-    nelemy = x.shape[0] - 1
-    nelem = nelemx * nelemy
-    nnx = x.shape[1]
-    e2vcg = np.zeros([4, nelem])
-    for j in range(nelemy):
-        for i in range(nelemx):
-            e2vcg[:, j * nelemx + i] = np.asarray(
-                [j * nnx + i, j * nnx + i + 1, (j + 1) * nnx + i, (j + 1) * nnx + i + 1]
-            )
-    return e2vcg.astype("int")
-
-
-def visualize2D(ax, x, y, u, colorbarPosition="vertical", colorlimit=None):
-    xdg0 = np.vstack([x.flatten(order="C"), y.flatten(order="C")])
-    udg0 = u.flatten(order="C")
-    idx = np.asarray([0, 1, 3, 2])
-    nelemx = x.shape[1] - 1
-    nelemy = x.shape[0] - 1
-    nelem = nelemx * nelemy
-    e2vcg0 = gen_e2vcg(x)
-    udg_ref = udg0[e2vcg0]
-    cmap = matplotlib.cm.coolwarm
-    polygon_list = []
-    for i in range(nelem):
-        polygon_ = Polygon(xdg0[:, e2vcg0[idx, i]].T)
-        polygon_list.append(polygon_)
-    polygon_ensemble = PatchCollection(polygon_list, cmap=cmap, alpha=1)
-    polygon_ensemble.set_edgecolor("face")
-    polygon_ensemble.set_array(np.mean(udg_ref, axis=0))
-    if colorlimit is None:
-        pass
-    else:
-        polygon_ensemble.set_clim(colorlimit)
-    ax.add_collection(polygon_ensemble)
-    ax.set_xlim(np.min(xdg0[0, :]), np.max(xdg0[0, :]))
-    ax.set_ylim(np.min(xdg0[1, :]), np.max(xdg0[1, :]))
-    cbar = plt.colorbar(polygon_ensemble, orientation=colorbarPosition)
-    return ax, cbar
-
-
-def dfdx(f, dydeta, dydxi, Jinv, h=0.01):
-    dfdxi_internal = (
-        (
-            -f[:, :, :, 4:]
-            + 8 * f[:, :, :, 3:-1]
-            - 8 * f[:, :, :, 1:-3]
-            + f[:, :, :, 0:-4]
-        )
-        / 12
-        / h
-    )
-    dfdxi_left = (
-        (
-            -11 * f[:, :, :, 0:-3]
-            + 18 * f[:, :, :, 1:-2]
-            - 9 * f[:, :, :, 2:-1]
-            + 2 * f[:, :, :, 3:]
-        )
-        / 6
-        / h
-    )
-    dfdxi_right = (
-        (
-            11 * f[:, :, :, 3:]
-            - 18 * f[:, :, :, 2:-1]
-            + 9 * f[:, :, :, 1:-2]
-            - 2 * f[:, :, :, 0:-3]
-        )
-        / 6
-        / h
-    )
-    dfdxi = paddle.concat(
-        (dfdxi_left[:, :, :, 0:2], dfdxi_internal, dfdxi_right[:, :, :, -2:]), 3
-    )
-    dfdeta_internal = (
-        (
-            -f[:, :, 4:, :]
-            + 8 * f[:, :, 3:-1, :]
-            - 8 * f[:, :, 1:-3, :]
-            + f[:, :, 0:-4, :]
-        )
-        / 12
-        / h
-    )
-    dfdeta_low = (
-        (
-            -11 * f[:, :, 0:-3, :]
-            + 18 * f[:, :, 1:-2, :]
-            - 9 * f[:, :, 2:-1, :]
-            + 2 * f[:, :, 3:, :]
-        )
-        / 6
-        / h
-    )
-    dfdeta_up = (
-        (
-            11 * f[:, :, 3:, :]
-            - 18 * f[:, :, 2:-1, :]
-            + 9 * f[:, :, 1:-2, :]
-            - 2 * f[:, :, 0:-3, :]
-        )
-        / 6
-        / h
-    )
-    dfdeta = paddle.concat(
-        (dfdeta_low[:, :, 0:2, :], dfdeta_internal, dfdeta_up[:, :, -2:, :]), 2
-    )
-    dfdx = Jinv * (dfdxi * dydeta - dfdeta * dydxi)
-    return dfdx
-
-
-def dfdy(f, dxdxi, dxdeta, Jinv, h=0.01):
-    dfdxi_internal = (
-        (
-            -f[:, :, :, 4:]
-            + 8 * f[:, :, :, 3:-1]
-            - 8 * f[:, :, :, 1:-3]
-            + f[:, :, :, 0:-4]
-        )
-        / 12
-        / h
-    )
-    dfdxi_left = (
-        (
-            -11 * f[:, :, :, 0:-3]
-            + 18 * f[:, :, :, 1:-2]
-            - 9 * f[:, :, :, 2:-1]
-            + 2 * f[:, :, :, 3:]
-        )
-        / 6
-        / h
-    )
-    dfdxi_right = (
-        (
-            11 * f[:, :, :, 3:]
-            - 18 * f[:, :, :, 2:-1]
-            + 9 * f[:, :, :, 1:-2]
-            - 2 * f[:, :, :, 0:-3]
-        )
-        / 6
-        / h
-    )
-    dfdxi = paddle.concat(
-        (dfdxi_left[:, :, :, 0:2], dfdxi_internal, dfdxi_right[:, :, :, -2:]), 3
-    )
-
-    dfdeta_internal = (
-        (
-            -f[:, :, 4:, :]
-            + 8 * f[:, :, 3:-1, :]
-            - 8 * f[:, :, 1:-3, :]
-            + f[:, :, 0:-4, :]
-        )
-        / 12
-        / h
-    )
-    dfdeta_low = (
-        (
-            -11 * f[:, :, 0:-3, :]
-            + 18 * f[:, :, 1:-2, :]
-            - 9 * f[:, :, 2:-1, :]
-            + 2 * f[:, :, 3:, :]
-        )
-        / 6
-        / h
-    )
-    dfdeta_up = (
-        (
-            11 * f[:, :, 3:, :]
-            - 18 * f[:, :, 2:-1, :]
-            + 9 * f[:, :, 1:-2, :]
-            - 2 * f[:, :, 0:-3, :]
-        )
-        / 6
-        / h
-    )
-    dfdeta = paddle.concat(
-        (dfdeta_low[:, :, 0:2, :], dfdeta_internal, dfdeta_up[:, :, -2:, :]), 2
-    )
-    dfdy = Jinv * (dfdeta * dxdxi - dfdxi * dxdeta)
-    return dfdy
-
-
-def tikzplotlib_fix_ncols(obj):
-    """
-    workaround for matplotlib 3.6 renamed legend's _ncol to _ncols, which breaks tikzplotlib
-    """
-    if hasattr(obj, "_ncols"):
-        obj._ncol = obj._ncols
-    for child in obj.get_children():
-        tikzplotlib_fix_ncols(child)
 
 
 def train(cfg: DictConfig):
@@ -235,7 +30,7 @@ def train(cfg: DictConfig):
     pad_singleside = cfg.MODEL.pad_singleside
     model = ppsci.arch.USCNN(**cfg.MODEL)
 
-    optimizer = ppsci.optimizer.Adam(cfg.TRAIN.lr)(model)
+    optimizer = ppsci.optimizer.Adam(cfg.TRAIN.learning_rate)(model)
 
     sup_constraint_mres = ppsci.constraint.SupervisedConstraint(
         {
@@ -250,7 +45,7 @@ def train(cfg: DictConfig):
                     "dydetas": dydetas,
                 },
             },
-            "batch_size": 1,
+            "batch_size": cfg.TRAIN.batch_size,
             "iters_per_epoch": coords.shape[0],
             "num_workers": 0,
         },
@@ -309,9 +104,8 @@ def evaluate(cfg: DictConfig):
     OFV_sb = paddle.to_tensor(data["OFV_sb"])
 
     ## create model
-    pad_singleside = 1
+    pad_singleside = cfg.MODEL.pad_singleside
     model = ppsci.arch.USCNN(**cfg.MODEL)
-    model.register_output_transform(None)
     solver = ppsci.solver.Solver(
         model,
         pretrained_model_path=cfg.EVAL.pretrained_model_path,  ### the path of the model
